% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gps.stops.matcher.R
\name{disambiguate.matched.stop}
\alias{disambiguate.matched.stop}
\title{Disambiguates matched-stop locations}
\usage{
disambiguate.matched.stop(location.row, initial.stops.sequence,
  trip.locations.df, trips.stops.df)
}
\arguments{
\item{location.row}{matched location row}

\item{initial.stops.sequence}{vector with matched stops}

\item{trip.locations.df}{data frame with GPS locations}

\item{trips.stops.df}{data frame with stops points}
}
\value{
best-matching point
}
\description{
This function analyzes points which matched the same stop to find which one best fits the line trip
}
\examples{

}

