% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gps.stops.matcher.R
\name{find.trip.initial.stop.seq}
\alias{find.trip.initial.stop.seq}
\title{Returns initial stop sequence for GPS bus trajectory}
\usage{
find.trip.initial.stop.seq(trajectory.location.data, line.stops.df,
  verbose = FALSE)
}
\arguments{
\item{trajectory.location.data}{data frame with GPS locations for a bus trajectory}

\item{line.stops.df}{line stops locations data frame}

\item{verbose}{if TRUE then debugging logs are printed during processing}
}
\value{
first stop-matched point in trajectory
}
\description{
This function compares bus trajectory points to line stops points to find the initial point of the trajectory. If the initial point matches more than one stop, a disambiguation process is performed.
}
\examples{

}

