% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gps.stops.matcher.R
\name{get.line.longest.trip}
\alias{get.line.longest.trip}
\title{Returns longest trip among line trips}
\usage{
get.line.longest.trip(line, stops.df)
}
\arguments{
\item{line}{line code}

\item{stops.df}{all lines stops data frame}
}
\value{
trip with the greatest number of stops among line trips
}
\description{
Returns longest trip among line trips
}
\examples{

}

