% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gps.stops.matcher.R
\name{get.stop.location.dist}
\alias{get.stop.location.dist}
\title{Returns distance between stop and GPS location points}
\usage{
get.stop.location.dist(locations.df, stops.df, stop.row, location.row)
}
\arguments{
\item{locations.df}{GPS locations data frame}

\item{stops.df}{stops locations data frame}

\item{stop.row}{row number of stop in stops.df to be analyzed}

\item{location.row}{row number of location in locations.df to be analyzed}
}
\value{
distance between stop at row stop.row and location at row location.row
}
\description{
Returns distance between stop and GPS location points
}
\examples{

}

