% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gps.stops.matcher.R
\name{get.trip.number}
\alias{get.trip.number}
\title{Returns point trip number}
\usage{
get.trip.number(point.timestamp, trips.inits.df)
}
\arguments{
\item{point.timestamp}{GPS point timestamp}

\item{trips.inits.df}{trips initial point (longitude,latitude,timestamp) dataframe}
}
\value{
point trip number
}
\description{
Matches point to trip according to point timestamp
}
\examples{

}

