% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gps.stops.matcher.R
\name{get.trips.initial.points}
\alias{get.trips.initial.points}
\title{Returns trips initial points}
\usage{
get.trips.initial.points(locations.df, initial.pt.df, thresh = 50)
}
\arguments{
\item{locations.df}{GPS locations data frame}

\item{initial.pt.df}{initial point (longitude,latitude) dataframe}

\item{thresh}{point 2 dataframe with latitude,longitude}

\item{threshold}{maximum distance for points matching}
}
\value{
distance between point 1 and point 2
}
\description{
Returns trips initial points
}
\examples{
locations <- data.frame(51.24,20.53)
init.pt <- data.frame(59.39,29.53)
get.trips.initial.points(locations,init.pt,30)

}

