% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gps.stops.matcher.R
\name{match.bus.locations.stops}
\alias{match.bus.locations.stops}
\title{Matches stops to GPS locations for a bus}
\usage{
match.bus.locations.stops(bus.locations.df, line.stops.df, verbose = FALSE)
}
\arguments{
\item{bus.locations.df}{data frame with GPS locations for a single bus}

\item{line.stops.df}{line stops locations data frame}

\item{verbose}{if TRUE then debugging logs are printed during processing}
}
\value{
data frame with matched stops and GPS locations data
}
\description{
This function splits bus GPS locations by trip and finds trips initial point to call match.trip.locations.stops
}
\examples{

}

