% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gps.stops.matcher.R
\name{match.line.locations.stops}
\alias{match.line.locations.stops}
\title{Matches stops to GPS locations for a line}
\usage{
match.line.locations.stops(line.location.data, stops.df, verbose = FALSE)
}
\arguments{
\item{line.location.data}{data frame with GPS locations for a single line}

\item{stops.df}{stops locations data frame}

\item{verbose}{if TRUE then debugging logs are printed during processing}
}
\value{
data frame with matched stops and GPS locations data
}
\description{
This function splits line GPS locations by bus and filters line stops to call match.bus.locations.stops. Finally, matched trips with size shorter than 1/3 of longest trip are removed and rows are sorted by bus code, trip number and timestamp.
}
\examples{

}

