% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gps.stops.matcher.R
\name{points.match}
\alias{points.match}
\title{Checks if the location of two geographical points match}
\usage{
\method{points}{match}(p1.lat, p1.lon, p2.df, threshold)
}
\arguments{
\item{p1.lat}{point 1 latitude}

\item{p1.lon}{point 1 longitude}

\item{p2.df}{point 2 dataframe with latitude,longitude}

\item{threshold}{maximum distance for matching}
}
\value{
distance between point 1 and point 2
}
\description{
Checks if the location of two geographical points match
}
\examples{
p2 <- data.frame(59.39,29.53)
points.match(51.09,25.94,p2,50)

}

