% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gps.stops.matcher.R
\name{prepare.gps.data}
\alias{prepare.gps.data}
\title{Reads and pre-processes GPS data}
\usage{
prepare.gps.data(bus.gps.csv.file.path)
}
\arguments{
\item{bus.gps.csv.file.path}{GPS data csv file path}
}
\value{
data frame with GPS data
}
\description{
This function reads GPS data from a csv file and parses the timestamps into POSIXct date-time objects
}
\examples{

}

