% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gps.stops.matcher.R
\name{prepare.stops.data}
\alias{prepare.stops.data}
\title{Reads and pre-processes stops data}
\usage{
prepare.stops.data(gtfs.folder.path)
}
\arguments{
\item{gtfs.folder.path}{GTFS folder path}
}
\value{
data frame with stops data
}
\description{
This function reads GTFS data from a csv file, assembles the necessary information and organizes it into a data frame for simpler usage.
}
\examples{

}

